refactor(store): use index and combine types to types.ts
