+++ live/src-vuepress/theme/store/types.js
@@ -28,3 +28,33 @@ const types = [
  **/
 export const MUTATION_TYPE = types.reduce((result, type) => ({ ...result, [type]: type }), {});
 export default MUTATION_TYPE;
+/** Initial state of the Vuex store.
+ */
+export const defaultState = {
+    blog: {
+        title: null,
+        description: null,
+        base: null,
+    },
+    header: {
+        showCover: false,
+        coverImage: null,
+        title: null,
+        description: null,
+    },
+    nav: [],
+    params: {},
+    current: {},
+    route: {},
+    index: [],
+    loading: true,
+    search: "",
+    sidebarOpen: false,
+    type: null,
+    posts: [],
+    footer: [],
+    social: [],
+    author: null,
+    allTags: new Set(),
+    allCategories: new Set(),
+};
